server와 api_gateway의 app.js에서 Express 라우터 분리
  - Express 라우터를 별도 파일로 분리함
    - server에 routes 디렉터리를 생성하고 auth.js, index.js, main.js 파일들 추가
    - api_gateway에 routes 디렉터리를 생성하고 api.js 파일 추가

  - server에 walkDataHandler.js, weatherHandler.js 추가하여 app.js 모듈화

  - api_gateway에 schedule.js파일 생성하여 app.js 모듈화
