+++ .github/workflows/CI_CD.yml
@@ -94,7 +94,13 @@ jobs:
           docker tag nginx:1.25 ${{secrets.DOCKER_REPO}}/nginx:1.25
           docker push ${{secrets.DOCKER_REPO}}/nginx:1.25
   Deploy_to_server:
-    needs: [ build_and_push_server, build_and_push_api_gateway, build_and_push_mysql, build_and_push_nginx]
+    needs:
+      [
+        build_and_push_server,
+        build_and_push_api_gateway,
+        build_and_push_mysql,
+        build_and_push_nginx,
+      ]
     runs-on: ubuntu-latest
     steps:
       - name: executing remote ssh commands using password
